split up AzureContainer in multiple files
