+++ Arius/Arius.Core.Tests/UnitTests/ChunkBlobAndChunkBlobEntryTests.cs
@@ -5,6 +5,7 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Arius.Core.Repositories.BlobRepository;
 
 namespace Arius.Core.Tests.UnitTests;
