+++ Arius/Arius.Core.Tests/UnitTests/ChunkListTests.BinariesTests.cs
@@ -7,6 +7,7 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Arius.Core.Repositories.BlobRepository;
 
 namespace Arius.Core.Tests.UnitTests;
