+++ Arius/Arius.Core/Repositories/Repository.cs
@@ -9,6 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Arius.Core.Repositories.BlobRepository;
 
 namespace Arius.Core.Repositories;
