+++ Arius/Arius.Core/Repositories/RepositoryBuilder.StateDbContextFactory.cs
@@ -9,6 +9,7 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Arius.Core.Repositories.BlobRepository;
 
 namespace Arius.Core.Repositories;
