+++ Arius/Arius.Core/Repositories/RepositoryBuilder.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.Logging;
 using System;
 using System.Net;
 using System.Threading.Tasks;
+using Arius.Core.Repositories.BlobRepository;
 
 namespace Arius.Core.Repositories;
